EXP-84. Update TopNextSongs to remove Java to Scala type conversions.
This commit modifies TopNextSongs so that it no longer needs a
conversion from Scala lists to Java lists before packing the list into
an AvroRecord. This amounts to just removing the method and call that
performs the conversion.

This change is dependent on EXP-85, which removes a bug in how Avro
records are packed.

Reviewed by: juliet